+++ Gemfile
@@ -2,6 +2,7 @@
 
 source 'http://rubygems.org'
 
+gem 'active_support'
 gem 'fhir_client'
 gem 'fhir_models'
 gem 'jsonpath'
