+++ Rakefile
@@ -2,4 +2,4 @@ begin
   require 'rspec/core/rake_task'
   RSpec::Core::RakeTask.new(:spec)
   task :default => :spec
-end 
+end
