+++ bin/testscript_engine
@@ -1,0 +1,4 @@
+#!/usr/bin/env ruby
+
+require 'testscript_engine'
+puts TestScriptEngine.hi
