+++ lib/testscript_engine/message_handler.rb
