+++ lib/testscript_engine/operation.rb
