+++ lib/testscript_engine/testreport_handler.rb
