+++ spec/assertions_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../lib/assertion'
+require_relative '../lib/testscript_engine/assertion'
 require 'fhir_models'
 require 'fhir_client'
