--- test_script_engine.gemspec
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-Gem::Specification.new do |s|
-  s.name = 'test_script_engine'
-  s.version = '0.0.0'
-  s.summary = 'A collection of tools for utilizing the FHIR TestScript Resource.'
-  s.author = 'John Fraser'
-  s.email = 'jfraser@mitre.org'
-  s.license = 'MIT'
-  s.files = Dir.glob('{lib}/*')
-end
